ml/digest-score/requirements.txt
psycopg2-binary
pandas
scikit-learn
joblib
google-cloud-storage
flask
pydantic
sklearn2pmml
sqlalchemy
pyarrow
prometheus_client
treeinterpreter
xgboost==2.1.0
